Bugs resolved.
There were two errors, one when trying to login, and another when detecting an active account. In the case of login, I applied a logic where it is only possible to log in if the account is not logged in, in detecting the account the error was that I did not set Loading to false when checking if there was an active account so the logic did not work.